fix: rename GetIdentityQuotasByAddress to GetIdentity
