+++ Modules/Quotas/src/Quotas.Application/Identities/Queries/GetIdentity/Handler.cs
@@ -3,8 +3,8 @@ using Backbone.Modules.Quotas.Domain.Aggregates.Identities;
 using Enmeshed.BuildingBlocks.Application.Abstractions.Exceptions;
 using MediatR;
 
-namespace Backbone.Modules.Quotas.Application.Identities.Queries.GetIdentityQuotasByAddress;
-public class Handler : IRequestHandler<GetIdentityQuotasByAddressQuery, GetIdentityQuotasByAddressResponse>
+namespace Backbone.Modules.Quotas.Application.Identities.Queries.GetIdentity;
+public class Handler : IRequestHandler<GetIdentityQuery, GetIdentityResponse>
 {
     private readonly IIdentitiesRepository _identitiesRepository;
     private readonly IMetricsRepository _metricsRepository;
@@ -15,13 +15,13 @@ public class Handler : IRequestHandler<GetIdentityQuotasByAddressQuery, GetIdent
         _metricsRepository = metricsRepository;
     }
 
-    public async Task<GetIdentityQuotasByAddressResponse> Handle(GetIdentityQuotasByAddressQuery request, CancellationToken cancellationToken)
+    public async Task<GetIdentityResponse> Handle(GetIdentityQuery request, CancellationToken cancellationToken)
     {
         var identity = await _identitiesRepository.Find(request.Address, cancellationToken) ?? throw new NotFoundException(nameof(Identity));
 
         var metricsKeys = identity.TierQuotas.Select(q => q.MetricKey).Union(identity.IndividualQuotas.Select(q => q.MetricKey));
         var metrics = await _metricsRepository.FindAllWithKeys(metricsKeys, cancellationToken);
 
-        return new GetIdentityQuotasByAddressResponse(identity.Address, identity.TierQuotas, identity.IndividualQuotas, metrics);
+        return new GetIdentityResponse(identity.Address, identity.TierQuotas, identity.IndividualQuotas, metrics);
     }
 }
